Linter Fixes
Code now matches closer to some kind of standard.

+++ Game/src/attack_states/moves/Guillotine.java
@@ -9,7 +9,7 @@ public class Guillotine extends Move {
         if (ourselves.getLevel() < opponent.getLevel() || ourselves.getCurSPD() < opponent.getCurSPD())
             return 0;
         else
-            return ourselves.getLevel() - opponent.getLevel() + 30;
+            return (double) (ourselves.getLevel() - opponent.getLevel() + 30);
     }
 
     @Override
